Update provisioning scripts

+++ app/libs/bugm_host.rb
@@ -30,6 +30,7 @@ class BugmHost
     end
 
     def reset_influx
+      return unless File.exist?("/etc/influxdb/influxdb.conf")
       InfluxStats.delete_database("bugm_stats")
       InfluxStats.create_database("bugm_stats")
     end
